Fixed subtle bug in PHAL::SDirichlet
As explained in the comments in the Def file, we have to be careful
when const-casting a const Thyra_Vector. In particular, if the Thyra
vector is indeed of type Thyra::TpetraVector, then it stores a
Teuchos::ConstNonconstObjectContainer for the Tpetra_Vector RCP.
If (as it is most likely) the RCP<const Thyra_Vector> was created
const, the const-nonconst container will not allow a call to the
nonconst rcp. See this example:

  RCP<const Thyra_Vector> v (new Thyra_TpetraVector(...));
  auto non_const_v = Teuchos::rcp_const_cast<Thyra_Vector>(v);

  auto tpetra_v = Teuchos::rcp_dynamic_cast<Thyra_TpetraVector>(non_const_v);

So far so good. However, if you try to call tpetra_v->getTpetraVector(),
teuchos will throw, because the const-nonconst container was created
from an RCP<const T>, so you cannot modify the content.